requirements.txt
requests
dropbox
pyyaml
crontab